HZN-1210: Add initiator flag to flow document and compare ip addresses correctly
