+++ features/telemetry/adapters/netflow/src/main/java/org/opennms/netmgt/telemetry/adapters/netflow/Netflow5Adapter.java
@@ -28,8 +28,8 @@
 
 package org.opennms.netmgt.telemetry.adapters.netflow;
 
+import java.math.BigInteger;
 import java.net.InetAddress;
-import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
@@ -46,7 +46,6 @@ import org.opennms.netmgt.flows.api.NetflowDocument;
 import org.opennms.netmgt.flows.api.NodeInfo;
 import org.opennms.netmgt.flows.api.PersistenceException;
 import org.opennms.netmgt.flows.classification.ClassificationEngine;
-import org.opennms.netmgt.flows.classification.DefaultClassificationEngine;
 import org.opennms.netmgt.model.OnmsNode;
 import org.opennms.netmgt.telemetry.adapters.api.Adapter;
 import org.opennms.netmgt.telemetry.adapters.api.TelemetryMessage;
@@ -296,6 +295,7 @@ public class Netflow5Adapter implements Adapter {
                 // Metadata from message
                 document.setExporterAddress(sourceAddress);
                 document.setLocation(location);
+                document.setInitiator(isInitiator(document));
 
                 // Node data
                 getNodeInfoFromCache(location, sourceAddress).ifPresent(node -> document.setExporterNodeInfo(node));
@@ -353,5 +353,19 @@ public class Netflow5Adapter implements Adapter {
                 LOG.error("Flow {} could not be persisted. Reason: {}", failedItem.getItem(), failedItem.getCause().getMessage(), failedItem.getCause());
             });
         }
+    }
+
+    // Determine if the provided flow is the initiator.
+    // Yes, this may not be 100% accurate, but is a very easy way of defining the direction of the flow in most cases.
+    protected static boolean isInitiator(NetflowDocument document) {
+        if (document.getSourcePort()  > document.getDestPort()) {
+            return true;
+        } else if (document.getSourcePort() == document.getDestPort()) {
+            // Tie breaker
+            final BigInteger sourceAddressAsInt = InetAddressUtils.toInteger(InetAddressUtils.addr(document.getIpv4SourceAddress()));
+            final BigInteger destAddressAsInt = InetAddressUtils.toInteger(InetAddressUtils.addr(document.getIpv4DestAddress()));
+            return sourceAddressAsInt.compareTo(destAddressAsInt) > 0;
+        }
+        return false;
     }
 }
